cambios en logica

+++ servidorK/src/main/kotlin/com/example/data/entities/UsersDAO.kt
@@ -38,7 +38,7 @@ class UsersDAO(id: EntityID<Long>) : LongEntity(id) {
 
     fun toUser():User{
         return User(
-            id.value.toInt(),
+            id.value,
             nombre,
             apellido,
             rol,
